js: Promote getting AppSKey to JoinServer
